Endrer barnetillegg dagsats type til integer

+++ ytelse-ung/src/main/java/no/nav/ung/sak/ytelse/ung/beregning/barnetillegg/Barnetillegg.java
@@ -1,8 +1,6 @@
 package no.nav.ung.sak.ytelse.ung.beregning.barnetillegg;
 
-import java.math.BigDecimal;
-
-public record Barnetillegg(BigDecimal dagsats, int antallBarn) {
+public record Barnetillegg(int dagsats, int antallBarn) {
 
     @Override
     public String toString() {
